pyproject.toml
[tool.poetry]
name = "fastapimongodbcloudtest"
version = "0.1.0"
description = ""
authors = ["nurai5 <nursultanaibekuulu1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
beautifulsoup4 = "^4.12.3"
httpx = "^0.26.0"
python-dotenv = "^1.0.1"
pymongo = "^4.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
